Fixes showLocksRequest parametrization and checkstyle errors
